Removing a user from a role also makes schedules using that user invalid

+++ Backend/Schichtpilot/Interfaces/IJobRoleService.cs
@@ -13,9 +13,9 @@ public interface IJobRoleService
     
     Task RemoveDependenciesToJobRoleAsync(int jobRoleId, int dependencyId);
     
-    Task AddUsersToJobRoleAsync(int id, List<int> userIds);
+    Task AddUsersToJobRoleAsync(int id, List<long> userIds);
     
-    Task RemoveUsersFromJobRoleAsync(int id, List<int> userIds);
+    Task RemoveUsersFromJobRoleAsync(int id, List<long> userIds);
     
     Task<JobRoleDto> GetJobRoleAsync(int id);
